+++ src/AUnitVerbose.h
@@ -35,6 +35,15 @@ SOFTWARE.
 #ifndef AUNIT_AUNIT_VERBOSE_H
 #define AUNIT_AUNIT_VERBOSE_H
 
+// Blacklist boards using new Arduino API due to incompatibilities. This
+// currently includes all megaAVR boards and SAMD21 boards using arduino::samd
+// >= 1.8.10. Boards using arduino:samd <= 1.8.9 or SparkFun:samd are fine.
+#if defined(ARDUINO_ARCH_MEGAAVR)
+#error MegaAVR not supported https://github.com/bxparks/AUnit/issues/56
+#elif defined(ARDUINO_ARCH_SAMD) && defined(ARDUINO_API_VERSION)
+#error SAMD21 with arduino:samd >= 1.8.10 not supported, see https://github.com/bxparks/AUnit/issues/66
+#endif
+
 #include "aunit/print64.h"
 #include "aunit/Verbosity.h"
 #include "aunit/Compare.h"
@@ -50,7 +59,7 @@ SOFTWARE.
 #include "aunit/TestMacros.h"
 
 // Version format: xxyyzz == "xx.yy.zz"
-#define AUNIT_VERSION 10401
-#define AUNIT_VERSION_STRING "1.4.1"
+#define AUNIT_VERSION 10500
+#define AUNIT_VERSION_STRING "1.5"
 
 #endif