+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # Changelog
 
 * Unreleased
+* 1.5 (2021-01-18)
+    * Add support for STM32duino.
+    * Blacklist megaAVR boards due to incompatibility with new Arduino API.
+      See Issue #56 (https://github.com/bxparks/AUnit/issues/56).
+    * Blacklist SAMD21 boards using `arduino:samd` core version >= 1.8.10
+      due to incompatibility with new Arduino API. See Issue #66
+      (https://github.com/bxparks/AUnit/issues/66).
+    * Enable `F()` for ESP8266 since
+      https://github.com/esp8266/Arduino/issues/3369 seems to have been fixed
+      some time ago.
 * 1.4.1 (2020-11-12)
     * Add Table of Contents to README.md to help navigation.
     * No functional change in this release.